requirements.txt
requests>=2.31.0
kafka-python>=2.0.2
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
pyspark==3.5.0
pymongo>=4.5.0
setuptools>=65.0.0
py4j>=0.10.9.7
findspark>=2.0.1
